in post check all methods, user logged in or not

+++ controllers/PostController.php
@@ -39,12 +39,10 @@ class PostController extends Controller
      */
     public function actionIndex()
     {
-        // Check if user is logged in
         if (Yii::$app->user->isGuest) {
             // Redirect to login page
             return $this->redirect(['site/login']);
         }
-
         $dataProvider = new ActiveDataProvider([
             'query' => Post::find(),
             'sort' => [
@@ -72,6 +70,10 @@ class PostController extends Controller
      */
     public function actionView($id)
     {
+        if (Yii::$app->user->isGuest) {
+            // Redirect to login page
+            return $this->redirect(['site/login']);
+        }
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
@@ -84,6 +86,10 @@ class PostController extends Controller
      */
     public function actionCreate()
     {
+        if (Yii::$app->user->isGuest) {
+            // Redirect to login page
+            return $this->redirect(['site/login']);
+        }
         $model = new Post();
 
         if ($this->request->isPost) {
@@ -108,6 +114,10 @@ class PostController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (Yii::$app->user->isGuest) {
+            // Redirect to login page
+            return $this->redirect(['site/login']);
+        }
         $model = $this->findModel($id);
 
         if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
@@ -128,6 +138,10 @@ class PostController extends Controller
      */
     public function actionDelete($id)
     {
+        if (Yii::$app->user->isGuest) {
+            // Redirect to login page
+            return $this->redirect(['site/login']);
+        }
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
